pyproject.toml
[project]
name = "VirtualTruck"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Monitor Team", email = "monitor@monitor.com.br"},
]
dependencies = [
    "django>=5.1.2",
    "djangorestframework>=3.15.2",
    "django-simple-history>=3.7.0",
    "django-safedelete>=1.4.0",
    "drf-spectacular>=0.27.2",
    "setuptools>=75.1.0",
    "cpf-cnpj-validate>=1.4",
    "Pillow>=11.0.0",
    "python-magic>=0.4.27; sys_platform == \"linux\"",
    "django-cors-headers>=4.5.0",
    "pytest>=8.3.3",
    "pylint-django>=2.6.1",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.isort]
profile = "black"

[tool.pdm.scripts]
manage = "python src/manage.py"
makemigrations = "pdm run manage makemigrations"
migrate = "pdm run manage migrate"
dev = "pdm run manage runserver 0.0.0.0:8000"
test = "pytest --ignore __pypackages__"

[tool.pdm.dev-dependencies]
dev = [
    "pytest>=8.3.3",
    "pytest-django>=4.9.0",
    "colorama>=0.4.6",
    "model-bakery>=1.20.0",
    "pylint>=3.3.1",
    "isort>=5.13.2",
]
